Reuse training status indicator across model table and training viewer
