+++ client/src/components/train/ViewTraining.tsx
@@ -1,8 +1,8 @@
 import {getModelLogs} from 'lib/api/models';
 import fileDownload from 'js-file-download';
-import React, {useEffect, useRef, useState} from 'react';
+import React, {useEffect, useState} from 'react';
 import Model, {TrainingStatus} from 'types/Model';
-import TrainingStatusIndicator from './TrainingStatusIndicator';
+import TrainingStatusIndicator from 'components/train/TrainingStatusIndicator';
 
 type Props = {
   model: Model;
@@ -12,7 +12,6 @@ const LOGS_REFRESH_RATE = 2000;
 
 const ViewTraining: React.FC<Props> = ({model}) => {
   const [logs, setLogs] = useState<string[]>([]);
-  const ref = useRef<HTMLDivElement | null>(null);
 
   // Continuously pull model logs if it's training.
   useEffect(() => {
@@ -22,11 +21,6 @@ const ViewTraining: React.FC<Props> = ({model}) => {
         const blob = await response.blob();
         const logs = await blob.text();
         setLogs(logs.split('\n'));
-        if (ref.current) {
-          ref.current.scrollTo({
-            top: ref.current.scrollHeight + ref.current.offsetHeight,
-          });
-        }
       } else {
         const error = await response.text();
         console.error('Error fetching model logs: ', error);
@@ -57,14 +51,16 @@ const ViewTraining: React.FC<Props> = ({model}) => {
     <div className="mt-8 space-y-4">
       <div className="flex items-center justify-between">
         <p className="subtitle">Training Logs</p>
-        <TrainingStatusIndicator
-          status={model.status ?? TrainingStatus.Training}
-        />
+        <div className="flex space-x-2">
+          <p className="capitalize font-semibold">
+            {model.status ?? TrainingStatus.Training}
+          </p>
+          <TrainingStatusIndicator
+            status={model.status ?? TrainingStatus.Training}
+          />
+        </div>
       </div>
-      <div
-        ref={ref}
-        className="max-h-96 overflow-y-auto bg-slate-800 text-white font-mono text-xs rounded p-3"
-      >
+      <div className="max-h-96 overflow-y-auto bg-slate-800 text-white font-mono text-xs rounded p-3">
         {logs.map((log, index) => (
           <p key={index}>{log}</p>
         ))}
